server/requirements.txt
anytree==2.8.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
distlib==0.3.2
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
Lasagne==0.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
numpy==1.20.1
pandas==1.2.2
Pillow==8.2.0
pipenv==2021.5.29
psycopg2==2.8.6
pycodestyle==2.8.0
pylint==2.6.2
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-mnist==0.7
python-telegram-bot==5.3.0
pytz==2021.1
requests==2.25.1
scipy==1.7.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.2.0
toml==0.10.2
transitions==0.5.0
urllib3==1.26.4
virtualenv==20.4.7
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
